+++ components/common/type.ts
@@ -8,21 +8,6 @@ export type StepTemplateProps = {
   imageDesc?: string;
 };
 
-export type TabLayoutProps = {
-  title: string;
-  icon: string;
-  content: string;
-  tabs: string[];
-  children: React.ReactNode | React.ReactNode[];
-};
-
-export type TabMenuProps = {
-  tabTitle: string;
-  isSelected: boolean;
-  isPrevSelected: boolean;
-  isNextSelected: boolean;
-};
-
 export type DeliverButtonProps = {
   isActive?: boolean;
   onClick: () => void;
@@ -35,9 +20,3 @@ export type CustomInputProps = {
   placeholder: string;
   className?: string;
 } & UseControllerProps;
-
-export type PageDescriptionProps = {
-  icon: string;
-  title: string;
-  content: string;
-};
